cors/CORS_011.htm

<!DOCTYPE html>
<html>
    <head>
        <title>Verify simple response headers are exposed to getAllResponseHeaders() for XHR requests</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta description="Verify simple response headers are always exposed to getAllResponseHeaders() for XHR requests" />
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
        <link rel="help" href="http://dvcs.w3.org/hg/cors/raw-file/tip/Overview.html#handling-a-response-to-a-cross-origin-request"/>
        <script src="http://w3c-test.org/resources/testharness.js"></script>
        <script src="http://w3c-test.org/resources/testharnessreport.js"></script>
        <script src="SetCORSUrls.js"></script>
    </head>
    <body>
        <div id=log></div>
        <script type="text/javascript">
            var t = async_test("Verify simple response headers are exposed to getAllResponseHeaders() for XHR requests");
            var xhr = new XMLHttpRequest();
            if('withCredentials' in xhr)
            {
                var url = CORS_SERVER_URL + "?TestID=11"
                xhr.open("GET", url, true);
                xhr.onreadystatechange = function(e)
                {
                    if (xhr.readyState == 4)
                    {
                        // 'Content-type' is a simple response header and should always be exposed
                        // via getAllResponseHeaders()
                        t.step(function(){assert_true(xhr.getAllResponseHeaders().toLowerCase().indexOf('content-type') !== -1)});
                        t.done();
                    }
                }
                xhr.send();
            }
            else
            {
                t.step(function(){assert_unreached("[Cross-Origin XMLHttpRequests are not supported]")});
                t.done();
            }
        </script>
    </body>
</html>
